unsquashfs.h: widen blocks in struct inode
The file_size can be long long in size, and this can produce a blocks
count that exceeds the int size.

+++ squashfs-tools/unsquashfs.h
@@ -78,7 +78,7 @@ struct hash_table_entry {
 };
 
 struct inode {
-	int		blocks;
+	long long	blocks;
 	long long	block_start;
 	unsigned int	block_offset;
 	long long	data;
